Fixed destroy method so it will only try to destroy votes if votes exist

+++ app/controllers/works_controller.rb
@@ -93,7 +93,10 @@ class WorksController < ApplicationController
   def destroy
     Work.destroy(params[:id])
     votes = Vote.find_by_work_id(params[:id])
-    votes.destroy
+
+    if votes
+      votes.destroy
+    end
 
     redirect_to root_path
   end
